move some things to Jumpvalley.Game
The Jumpvalley.Game namespace isn't part of the core API.

This namespace was created so that code in this namespace can more easily receive changes that break backwards-compatibility without worrying about breaking backwards-compatibility for the core API.
