+++ src/Game/MeshSpinner.cs
@@ -2,7 +2,7 @@ using Godot;
 
 // refer to this article for naming conventions:
 // https://learn.microsoft.com/en-us/dotnet/csharp/fundamentals/coding-style/coding-conventions
-namespace Jumpvalley
+namespace Jumpvalley.Game
 {
     public partial class MeshSpinner : Node
     {
